--- REFACTORING_SUMMARY.md
@@ -1,92 +1,0 @@
-# Refactoring Summary
-
-## Overview
-Completed comprehensive refactoring to improve code maintainability without changing functionality.
-
-## Changes Made
-
-### 1. TypeScript Configuration Improvements
-- **File**: `tsconfig.json`
-- **Changes**: 
-  - Enabled `noImplicitReturns` for better type safety
-  - Enabled `noUnusedLocals` and `noUnusedParameters` to catch dead code
-  - Fixed unused parameter warnings by prefixing with underscore
-
-### 2. Extracted Phase Handlers from QueenBeeOrchestrator
-- **Original**: Single 824-line orchestrator class
-- **Refactored into**:
-  - `src/orchestrator/phases/phase1-analyzer.ts` - Initial analysis logic
-  - `src/orchestrator/phases/phase2-improver.ts` - Improvement logic
-  - `src/orchestrator/phases/phase3-finalizer.ts` - Finalization logic
-  - `src/orchestrator/phases/phase4-pattern-extractor.ts` - Pattern extraction
-  - `src/orchestrator/queen-bee-orchestrator-refactored.ts` - Simplified main orchestrator
-
-**Benefits**:
-- Single Responsibility: Each phase handler has one clear purpose
-- Better testability: Phases can be tested independently
-- Reduced complexity: Main orchestrator reduced from 824 to ~250 lines
-- Improved maintainability: Changes to one phase don't affect others
-
-### 3. Modularized Configuration
-- **Original**: Single 345-line `config.ts` file
-- **Refactored into**:
-  - `src/config/evaluation-criteria.ts` - Evaluation scoring criteria
-  - `src/config/model-config.ts` - LLM model configurations
-  - `src/config/app-config.ts` - Application settings
-  - `src/config/langfuse-config.ts` - Langfuse client setup
-  - `src/config/index.ts` - Central export point
-
-**Benefits**:
-- Separation of Concerns: Each config module handles one aspect
-- Easier to maintain: Find and modify specific configurations quickly
-- Better organization: Related settings grouped together
-- Reduced file size: Each file is focused and manageable
-
-### 4. Code Quality Improvements
-- Fixed TypeScript compilation warnings
-- Removed unused imports and variables
-- Added proper type annotations where missing
-- Improved code organization and structure
-
-## Metrics
-
-### Before Refactoring
-- Largest file: 824 lines (queen-bee-orchestrator.ts)
-- Config file: 345 lines
-- TypeScript strict mode: Disabled
-- Multiple `any` types throughout
-
-### After Refactoring
-- Largest new file: ~250 lines
-- Config split into 4 focused modules (50-100 lines each)
-- TypeScript stricter settings enabled
-- Reduced usage of `any` types
-
-## Test Results
-- **All 142 tests passing** ✅
-- No functionality changes
-- No breaking changes
-- Performance maintained
-
-## Key Principles Applied
-1. **Single Responsibility Principle**: Each module/class has one reason to change
-2. **DRY (Don't Repeat Yourself)**: Extracted common patterns
-3. **Separation of Concerns**: Split large files by domain
-4. **KISS (Keep It Simple)**: Simplified complex orchestration logic
-5. **Open/Closed Principle**: New phase handlers can be added without modifying existing ones
-
-## Next Steps (Recommended)
-1. Continue replacing `any` types with proper interfaces
-2. Add unit tests for new phase handler classes
-3. Consider extracting more common patterns from optimizers
-4. Enable full TypeScript strict mode gradually
-5. Add JSDoc comments to new modules for better documentation
-
-## Files Modified
-- `tsconfig.json` - Stricter TypeScript settings
-- `src/claude-code-hook.ts` - Fixed unused parameters
-- `src/utils/logger.ts` - Updated import paths
-- Created 9 new files for better organization
-
-## Conclusion
-The refactoring successfully improved code maintainability and organization without breaking any existing functionality. The codebase is now more modular, easier to understand, and better prepared for future enhancements.
