+++ src/config/evaluation-criteria.ts
@@ -15,65 +15,64 @@ export interface CriterionConfig {
 
 export const EVALUATION_CRITERIA: Record<EvaluationCriterion, CriterionConfig> = Object.freeze({
   clarity: {
-    weight: 1.5,
+    weight: 1.2,
     minScore: 0,
     maxScore: 10,
     description: 'How clear and unambiguous the prompt is',
     examples: ['Clear task objectives', 'Well-defined requirements'],
   },
-  specificity: {
-    weight: 1.3,
-    minScore: 0,
-    maxScore: 10,
-    description: 'Level of detail and specific requirements',
-    examples: ['Specific output format', 'Detailed constraints'],
-  },
-  context_sufficiency: {
-    weight: 1.2,
-    minScore: 0,
-    maxScore: 10,
-    description: 'Adequacy of background information provided',
-  },
   structure: {
-    weight: 1.0,
+    weight: 1.1,
     minScore: 0,
     maxScore: 10,
     description: 'Organization and logical flow of the prompt',
   },
-  feasibility: {
-    weight: 1.1,
-    minScore: 0,
-    maxScore: 10,
-    description: 'Whether the task is achievable as specified',
-  },
-  edge_case_handling: {
-    weight: 0.9,
-    minScore: 0,
-    maxScore: 10,
-    description: 'Consideration of edge cases and error scenarios',
-  },
-  output_format: {
-    weight: 0.8,
-    minScore: 0,
-    maxScore: 10,
-    description: 'Clear specification of desired output format',
-  },
-  examples_quality: {
+  examples: {
     weight: 1.0,
     minScore: 0,
     maxScore: 10,
     description: 'Quality and relevance of provided examples',
   },
-  language_quality: {
-    weight: 0.7,
+  chainOfThought: {
+    weight: 1.1,
     minScore: 0,
     maxScore: 10,
-    description: 'Grammar, spelling, and language clarity',
+    description: 'Step-by-step reasoning and thinking process',
   },
-  technique_appropriateness: {
+  techSpecificity: {
     weight: 1.2,
     minScore: 0,
     maxScore: 10,
-    description: 'Use of appropriate prompting techniques',
+    description: 'Technical details and specific requirements',
+  },
+  errorHandling: {
+    weight: 1.0,
+    minScore: 0,
+    maxScore: 10,
+    description: 'Consideration of edge cases and error scenarios',
+  },
+  performance: {
+    weight: 0.9,
+    minScore: 0,
+    maxScore: 10,
+    description: 'Performance requirements and optimization',
+  },
+  testing: {
+    weight: 0.9,
+    minScore: 0,
+    maxScore: 10,
+    description: 'Testing requirements and validation',
+  },
+  outputFormat: {
+    weight: 1.0,
+    minScore: 0,
+    maxScore: 10,
+    description: 'Clear specification of desired output format',
+  },
+  deployment: {
+    weight: 0.8,
+    minScore: 0,
+    maxScore: 10,
+    description: 'Deployment and production readiness',
   },
 });
